Fea/authen author

+++ server/src/api/controllers/authController.js
@@ -1,9 +1,25 @@
-const createError = require("http-errors");
+const createError = require('http-errors');
 const fs = require('fs-extra');
 
-const authData = require("../services/auth");
-const { hashing, compareHashing } = require("../utils/hash_utils");
+const authData = require('../services/auth');
+const { hashing, compareHashing } = require('../utils/hash_utils');
 const { modifyUserInfo } = require('../utils/response_modifiers');
+
+const getCurrentUser = async (req, res, next) => {
+    try {
+        const id = req.payload.id;
+        const currentUser = await authData.getCurrentUser(id);
+
+        delete currentUser.password;
+
+        return res.send({
+            ...currentUser,
+            image: currentUser.image && `${process.env.HOST_URL}/profile/${currentUser.image}`,
+        });
+    } catch (error) {
+        next(createError(error.message));
+    }
+};
 
 const createSellerAccount = async (req, res, next) => {
     try {
@@ -12,12 +28,12 @@ const createSellerAccount = async (req, res, next) => {
         data.password = await hashing(data.password);
 
         const created = await authData.createSellerAccount(data);
-        
-        const response = await modifyUserInfo(created)
+
+        const response = await modifyUserInfo(created);
 
         return res.send(response);
     } catch (error) {
-        next(createError(error.status, error.message));
+        next(createError(error.message));
     }
 };
 
@@ -25,30 +41,32 @@ const updateSeller = async (req, res, next) => {
     try {
         const id = req.payload.id;
         if (!id) {
-            return next(createError.InternalServerError('Cannot get id'))
+            return next(createError.InternalServerError('Cannot get id'));
         }
 
         const seller = await authData.readSellerById(id);
-        if (seller.image) {
-            await fs.remove(`${process.cwd()}/public/images/profile/${seller.image}`);
-        }
-        const updatedSeller = Object.assign(seller, {
-            ...req.body,
-            image: req.file.filename,
-        })
+        const updatedSeller = Object.assign(seller, req.body);
 
-        const response = await authData.updateSeller(updatedSeller)
+        if (req.file) {
+            if (seller.image) {
+                await fs.remove(`${process.cwd()}/public/images/profile/${seller.image}`);
+            }
+
+            updatedSeller.image = req.file.filename;
+        }
+
+        const response = await authData.updateSeller(updatedSeller);
 
         return res.send({
             data: {
                 ...response,
-                image: `${process.env.HOST_URL}/profile/${req.file.filename}`
-            }
-        })
+                image: `${process.env.HOST_URL}/profile/${response.image}`,
+            },
+        });
     } catch (error) {
-        next(createError(error.status, error.message));
+        next(createError(error.message));
     }
-}
+};
 
 const sellerLogin = async (req, res, next) => {
     try {
@@ -57,28 +75,30 @@ const sellerLogin = async (req, res, next) => {
         const seller = await authData.readOneSeller(email);
 
         if (!seller) {
-            return next(createError.NotFound("Email address is not exist"));
+            return next(createError.NotFound('Email address is not exist'));
         }
 
-        const isPasswordValid = await compareHashing(
-            password,
-            seller.password
-        );
+        const isPasswordValid = await compareHashing(password, seller.password);
 
         if (!isPasswordValid) {
-            return next(createError.Unauthorized("Incorrect password"));
+            return next(createError.Unauthorized('Incorrect password'));
         }
-        
-        const response = await modifyUserInfo(seller)
+
+        if (seller.image) {
+            seller.image = `${process.env.HOST_URL}/profile/${seller.image}`;
+        }
+
+        const response = await modifyUserInfo(seller);
 
         return res.send(response);
     } catch (error) {
-        next(createError(error.status, error.message));
+        next(createError(error.message));
     }
 };
 
 module.exports = {
     createSellerAccount,
     sellerLogin,
-    updateSeller
+    updateSeller,
+    getCurrentUser,
 };
